Refactor authentication service and update related routes
