+++ app/api/core/dto/v1/auth/__init__.py
@@ -35,6 +35,4 @@ class TenantSlugData(BaseDTO):
 
 
 class LoginResponseData(BaseDTO):
-    access_token: str = Field(..., description="JWT access token")
-    token_type: str = Field(default="bearer", description="Access token type")
-    expires_in: int = Field(..., description="Token expiry in seconds")
+    at: str = Field(..., description="JWT access token")
