+++ app/api/routes/v1/auth.py
@@ -29,7 +29,7 @@ async def login(
     session: PostgresSession,
     auth_service: AuthServiceDep,
 ) -> SuccessResponse[LoginResponseData]:
-    access_token = await auth_service.login_user(
+    access_token = await auth_service.login(
         session=session,
         email=credentials.email,
         password=credentials.password,
@@ -53,6 +53,7 @@ async def login(
 async def register(
     data: RegisterDTO,
     session: PostgresSession,
+    auth_service: AuthServiceDep,
     user_service: UserServiceDep,
     email_service: EmailServiceDep,
 ) -> CreatedResponse[RegisterCreatedData]:
@@ -62,7 +63,7 @@ async def register(
         password=data.password,
         restaurant_name=data.restaurant_name,
     )
-    activation = await email_service.create_activation_link(
+    activation = await auth_service.create_activation_link(
         session=session,
         email=user.email,
         user_id=user.id,
@@ -96,9 +97,9 @@ async def register(
     response_description="Tenant activated successfully",
 )
 async def activate(
-    activation_id: UUID, session: PostgresSession, email_service: EmailServiceDep
+    activation_id: UUID, session: PostgresSession, auth_service: AuthServiceDep
 ) -> SuccessResponse[TenantSlugData]:
-    tenant, already_activated = await email_service.activate_account(
+    tenant, already_activated = await auth_service.activate_account(
         session=session, activation_id=activation_id
     )
     return SuccessResponse(
@@ -120,9 +121,10 @@ async def activate(
 async def resend_activation(
     activation_id: UUID,
     session: PostgresSession,
+    auth_service: AuthServiceDep,
     email_service: EmailServiceDep,
 ) -> SuccessResponse[TenantSlugData]:
-    new_link, tenant = await email_service.resend_activation_link(
+    new_link, tenant = await auth_service.resend_activation_link(
         session=session, activation_id=activation_id
     )
     activation_url = f"{settings.FRONTEND_URL}/activate?activation_id={new_link.id}"
